+++ src/pages/Settings/LogOut/LogOutButton.js
@@ -1,8 +1,8 @@
 import { LogOutButtonStyle } from "./LogOutButton.sc";
 import LogoutIcon from "@mui/icons-material/Logout";
-import { zeeguuOrange } from "../../components/colors";
+import { zeeguuOrange } from "../../../components/colors";
 import { useContext } from "react";
-import { UserContext } from "../../contexts/UserContext";
+import { UserContext } from "../../../contexts/UserContext";
 
 export default function LogOutButton() {
   const { logoutMethod } = useContext(UserContext);
@@ -12,8 +12,7 @@ export default function LogOutButton() {
         logoutMethod();
       }}
     >
-      Log Out{" "}
-      <LogoutIcon className="navigationIcon" sx={{ color: zeeguuOrange }} />
+      Log Out <LogoutIcon className="navigationIcon" sx={{ color: zeeguuOrange }} />
     </LogOutButtonStyle>
   );
 }
