+++ src/pages/Settings/Pages/DeleteAccount.js
@@ -1,10 +1,10 @@
 import { useEffect } from "react";
-import DeleteAccountButton from "../DeleteAccount/DeleteAccountButton";
-import PreferencesPage from "../_pages_shared/PreferencesPage";
-import Main from "../_pages_shared/Main.sc";
-import SettingsPageHeader from "./settings_pages_shared/SettingsPageHeader";
-import strings from "../../i18n/definitions";
-import { setTitle } from "../../assorted/setTitle";
+import DeleteAccountButton from "../../DeleteAccount/DeleteAccountButton";
+import PreferencesPage from "../../_pages_shared/PreferencesPage";
+import Main from "../../_pages_shared/Main.sc";
+import SettingsPageHeader from "../SharedComponents/SettingsPageHeader";
+import strings from "../../../i18n/definitions";
+import { setTitle } from "../../../assorted/setTitle";
 
 export default function DeleteAccount() {
   useEffect(() => {
