+++ src/pages/Settings/Pages/ExerciseSchedulingPreferences.js
@@ -1,21 +1,21 @@
 import { Link, useHistory } from "react-router-dom";
 import { useContext, useEffect } from "react";
-import { UserContext } from "../../contexts/UserContext";
-import Button from "../_pages_shared/Button.sc";
-import ButtonContainer from "../_pages_shared/ButtonContainer.sc";
-import Form from "../_pages_shared/Form.sc";
-import FormSection from "../_pages_shared/FormSection.sc";
-import PreferencesPage from "../_pages_shared/PreferencesPage";
-import Main from "../_pages_shared/Main.sc";
-import SettingsPageHeader from "./settings_pages_shared/SettingsPageHeader";
-import strings from "../../i18n/definitions";
-import { setTitle } from "../../assorted/setTitle";
-import { APIContext } from "../../contexts/APIContext";
-import InputField from "../../components/InputField";
-import useFormField from "../../hooks/useFormField";
-import { PositiveIntegerValidator } from "../../utils/ValidatorRule/Validator";
-import validateRules from "../../assorted/validateRules";
-import { ExercisesCounterContext } from "../../exercises/ExercisesCounterContext";
+import { UserContext } from "../../../contexts/UserContext";
+import Button from "../../_pages_shared/Button.sc";
+import ButtonContainer from "../../_pages_shared/ButtonContainer.sc";
+import Form from "../../_pages_shared/Form.sc";
+import FormSection from "../../_pages_shared/FormSection.sc";
+import PreferencesPage from "../../_pages_shared/PreferencesPage";
+import Main from "../../_pages_shared/Main.sc";
+import SettingsPageHeader from "../SharedComponents/SettingsPageHeader";
+import strings from "../../../i18n/definitions";
+import { setTitle } from "../../../assorted/setTitle";
+import { APIContext } from "../../../contexts/APIContext";
+import InputField from "../../../components/InputField";
+import useFormField from "../../../hooks/useFormField";
+import { PositiveIntegerValidator } from "../../../utils/ValidatorRule/Validator";
+import validateRules from "../../../assorted/validateRules";
+import { ExercisesCounterContext } from "../../../exercises/ExercisesCounterContext";
 
 const PREF_KEY_MAX_WORDS_TO_SCHEDULE = "max_words_to_schedule";
