+++ src/pages/Settings/Pages/LanguageSettings.js
@@ -1,30 +1,30 @@
 import { useHistory } from "react-router-dom";
 import { useState, useEffect, useContext, useRef } from "react";
-import { SystemLanguagesContext } from "../../contexts/SystemLanguagesContext";
-import { APIContext } from "../../contexts/APIContext";
-import { UserContext } from "../../contexts/UserContext";
-import { saveSharedUserInfo } from "../../utils/cookies/userInfo";
-import { CEFR_LEVELS } from "../../assorted/cefrLevels";
-import { getUserCefrLevel } from "../../utils/misc/cefrHelpers";
-import strings from "../../i18n/definitions";
-import LocalStorage from "../../assorted/LocalStorage";
-import LoadingAnimation from "../../components/LoadingAnimation";
-import LanguageSelector from "../../components/LanguageSelector";
-import Button from "../_pages_shared/Button.sc";
-import ButtonContainer from "../_pages_shared/ButtonContainer.sc";
-import Form from "../_pages_shared/Form.sc";
-import FormSection from "../_pages_shared/FormSection.sc";
-import PreferencesPage from "../_pages_shared/PreferencesPage";
-import Main from "../_pages_shared/Main.sc";
-import FullWidthErrorMsg from "../../components/FullWidthErrorMsg.sc";
-import SettingsPageHeader from "./settings_pages_shared/SettingsPageHeader";
-import Selector from "../../components/Selector";
-import { setTitle } from "../../assorted/setTitle";
-import useFormField from "../../hooks/useFormField";
-import { NonEmptyValidator, Validator } from "../../utils/ValidatorRule/Validator";
-import useShadowRef from "../../hooks/useShadowRef";
-import { scrollToTop } from "../../utils/misc/scrollToTop";
-import validateRules from "../../assorted/validateRules";
+import { SystemLanguagesContext } from "../../../contexts/SystemLanguagesContext";
+import { APIContext } from "../../../contexts/APIContext";
+import { UserContext } from "../../../contexts/UserContext";
+import { saveSharedUserInfo } from "../../../utils/cookies/userInfo";
+import { CEFR_LEVELS } from "../../../assorted/cefrLevels";
+import { getUserCefrLevel } from "../../../utils/misc/cefrHelpers";
+import strings from "../../../i18n/definitions";
+import LocalStorage from "../../../assorted/LocalStorage";
+import LoadingAnimation from "../../../components/LoadingAnimation";
+import LanguageSelector from "../../../components/LanguageSelector";
+import Button from "../../_pages_shared/Button.sc";
+import ButtonContainer from "../../_pages_shared/ButtonContainer.sc";
+import Form from "../../_pages_shared/Form.sc";
+import FormSection from "../../_pages_shared/FormSection.sc";
+import PreferencesPage from "../../_pages_shared/PreferencesPage";
+import Main from "../../_pages_shared/Main.sc";
+import FullWidthErrorMsg from "../../../components/FullWidthErrorMsg.sc";
+import SettingsPageHeader from "../SharedComponents/SettingsPageHeader";
+import Selector from "../../../components/Selector";
+import { setTitle } from "../../../assorted/setTitle";
+import useFormField from "../../../hooks/useFormField";
+import { NonEmptyValidator, Validator } from "../../../utils/ValidatorRule/Validator";
+import useShadowRef from "../../../hooks/useShadowRef";
+import { scrollToTop } from "../../../utils/misc/scrollToTop";
+import validateRules from "../../../assorted/validateRules";
 
 export default function LanguageSettings() {
   const api = useContext(APIContext);
