+++ src/pages/Settings/Pages/MyClassrooms/LeaveClassroomModal.js
@@ -1,10 +1,10 @@
 import React from "react";
-import Modal from "../../../components/modal_shared/Modal";
-import Header from "../../../components/modal_shared/Header.sc";
-import Heading from "../../../components/modal_shared/Heading.sc";
-import Footer from "../../../components/modal_shared/Footer.sc";
-import ButtonContainer from "../../../components/modal_shared/ButtonContainer.sc";
-import Button from "../../_pages_shared/Button.sc";
+import Modal from "../../../../components/modal_shared/Modal";
+import Header from "../../../../components/modal_shared/Header.sc";
+import Heading from "../../../../components/modal_shared/Heading.sc";
+import Footer from "../../../../components/modal_shared/Footer.sc";
+import ButtonContainer from "../../../../components/modal_shared/ButtonContainer.sc";
+import Button from "../../../_pages_shared/Button.sc";
 
 import * as s from "./LeaveClassroomModal.sc";
 
@@ -19,14 +19,10 @@ export default function LeaveClassroomModal({
     handleCloseLeaveClassroomModal();
   }
   return (
-    <Modal
-      open={isLeaveClassroomModalOpen}
-      onClose={handleCloseLeaveClassroomModal}
-    >
+    <Modal open={isLeaveClassroomModalOpen} onClose={handleCloseLeaveClassroomModal}>
       <Header>
         <Heading>
-          You are about to leave the following classroom:{" "}
-          <s.ColorAccent>{currentClassroom.name}</s.ColorAccent>
+          You are about to leave the following classroom: <s.ColorAccent>{currentClassroom.name}</s.ColorAccent>
         </Heading>
       </Header>
       <Footer>
