+++ src/pages/Settings/Pages/MyClassrooms/LeaveClassroomModal.sc.js
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import { zeeguuDarkOrange } from "../../../components/colors";
+import { zeeguuDarkOrange } from "../../../../components/colors";
 
 const ColorAccent = styled.span`
   font-size: inherit;
