+++ src/pages/Settings/Pages/MyClassrooms/MyClassrooms.js
@@ -1,26 +1,26 @@
 import React from "react";
 import { useState, useEffect, useContext } from "react";
-import { UserContext } from "../../../contexts/UserContext";
-import Button from "../../_pages_shared/Button.sc";
-import ButtonContainer from "../../_pages_shared/ButtonContainer.sc";
-import InputField from "../../../components/InputField";
-import Form from "../../_pages_shared/Form.sc";
-import FormSection from "../../_pages_shared/FormSection.sc";
-import PreferencesPage from "../../_pages_shared/PreferencesPage";
-import Main from "../../_pages_shared/Main.sc";
-import FullWidthListContainer from "../../../components/FullWidthListContainer.sc";
-import FullWidthErrorMsg from "../../../components/FullWidthErrorMsg.sc";
-import SettingsPageHeader from "../settings_pages_shared/SettingsPageHeader";
-import strings from "../../../i18n/definitions";
-import LoadingAnimation from "../../../components/LoadingAnimation";
-import FullWidthListItem from "../../../components/FullWidthListItem";
+import { UserContext } from "../../../../contexts/UserContext";
+import Button from "../../../_pages_shared/Button.sc";
+import ButtonContainer from "../../../_pages_shared/ButtonContainer.sc";
+import InputField from "../../../../components/InputField";
+import Form from "../../../_pages_shared/Form.sc";
+import FormSection from "../../../_pages_shared/FormSection.sc";
+import PreferencesPage from "../../../_pages_shared/PreferencesPage";
+import Main from "../../../_pages_shared/Main.sc";
+import FullWidthListContainer from "../../../../components/FullWidthListContainer.sc";
+import FullWidthErrorMsg from "../../../../components/FullWidthErrorMsg.sc";
+import SettingsPageHeader from "../../SharedComponents/SettingsPageHeader";
+import strings from "../../../../i18n/definitions";
+import LoadingAnimation from "../../../../components/LoadingAnimation";
+import FullWidthListItem from "../../../../components/FullWidthListItem";
 import LeaveClassroomModal from "./LeaveClassroomModal";
-import useFormField from "../../../hooks/useFormField";
-import { NonEmptyValidator } from "../../../utils/ValidatorRule/Validator";
-import { setTitle } from "../../../assorted/setTitle";
-import { APIContext } from "../../../contexts/APIContext";
-import LocalStorage from "../../../assorted/LocalStorage";
-import { saveSharedUserInfo } from "../../../utils/cookies/userInfo";
+import useFormField from "../../../../hooks/useFormField";
+import { NonEmptyValidator } from "../../../../utils/ValidatorRule/Validator";
+import { setTitle } from "../../../../assorted/setTitle";
+import { APIContext } from "../../../../contexts/APIContext";
+import LocalStorage from "../../../../assorted/LocalStorage";
+import { saveSharedUserInfo } from "../../../../utils/cookies/userInfo";
 
 export default function MyClassrooms() {
   const api = useContext(APIContext);
