+++ src/pages/Settings/Pages/Settings.sc.js
