+++ src/pages/Settings/SharedComponents/SettingsPageHeader.js
@@ -1,7 +1,7 @@
-import BackArrow from "./BackArrow";
-import Header from "../../_pages_shared/Header";
-import Heading from "../../_pages_shared/Heading.sc";
-import { HeaderWrapper, BackArrowWrapper } from "../Settings.sc";
+import BackArrow from "./BackArrow.js";
+import Header from "../../_pages_shared/Header.js";
+import Heading from "../../_pages_shared/Heading.sc.js";
+import { HeaderWrapper, BackArrowWrapper } from "../Pages/Settings.sc.js";
 
 export default function SettingsPageHeader({ title, redirectLink, func, children }) {
   return (
