+++ src/pages/Settings/_SettingsRouter.js
@@ -1,15 +1,15 @@
 import { PrivateRoute } from "../../PrivateRoute";
 import { Switch } from "react-router-dom";
-import Settings from "./Settings";
-import ProfileDetails from "./ProfileDetails";
-import LanguageSettings from "./LanguageSettings";
-import FeedPreferences from "./FeedPreferences";
-import ExerciseTypePreferences from "./ExerciseTypePreferences";
-import MyClassrooms from "./MyClassrooms/MyClassrooms";
-import DeleteAccount from "./DeleteAccount";
-import ExerciseSchedulingPreferences from "./ExerciseSchedulingPreferences";
-import DisplayPreferences from "./DisplayPreferences";
-import Developer from "./Developer";
+import Settings from "./Pages/Settings";
+import ProfileDetails from "./Pages/ProfileDetails";
+import LanguageSettings from "./Pages/LanguageSettings";
+import FeedPreferences from "./Pages/FeedPreferences";
+import ExerciseTypePreferences from "./Pages/ExerciseTypePreferences";
+import MyClassrooms from "./Pages/MyClassrooms/MyClassrooms";
+import DeleteAccount from "./Pages/DeleteAccount";
+import ExerciseSchedulingPreferences from "./Pages/ExerciseSchedulingPreferences";
+import DisplayPreferences from "./Pages/DisplayPreferences";
+import Developer from "./Pages/Developer";
 
 export default function SettingsRouter({ setUser }) {
   return (
